Move screens and read organization from slug
