+++ apps/web/pages/en/organizations/[slug]/services/index.ts
@@ -2,6 +2,6 @@ import '@island.is/api/mocks'
 
 import withApollo from '@island.is/web/graphql/withApollo'
 import { withLocale } from '@island.is/web/i18n'
-import ServicesPage from '@island.is/web/screens/Syslumenn/Services'
+import ServicesPage from '@island.is/web/screens/Organization/Services'
 
-export default withApollo(withLocale('is')(ServicesPage))
+export default withApollo(withLocale('en')(ServicesPage))
