+++ apps/web/pages/stofnanir/[slug]/[subSlug]/index.ts
@@ -2,6 +2,6 @@ import '@island.is/api/mocks'
 
 import withApollo from '@island.is/web/graphql/withApollo'
 import { withLocale } from '@island.is/web/i18n'
-import subPage from '@island.is/web/screens/Syslumenn/SubPage'
+import subPage from '@island.is/web/screens/Organization/SubPage'
 
-export default withApollo(withLocale('en')(subPage))
+export default withApollo(withLocale('is')(subPage))
