+++ apps/web/pages/stofnanir/syslumenn/frett/[slug]/index.tsx
@@ -1,5 +1,5 @@
 import withApollo from '@island.is/web/graphql/withApollo'
 import { withLocale } from '@island.is/web/i18n'
-import newsItemScreen from '@island.is/web/screens/Syslumenn/NewsItem'
+import newsItemScreen from '@island.is/web/screens/Organization/NewsItem'
 
 export default withApollo(withLocale('is')(newsItemScreen))
