+++ apps/web/pages/stofnanir/syslumenn/frett/index.tsx
@@ -1,5 +1,5 @@
 import withApollo from '@island.is/web/graphql/withApollo'
 import { withLocale } from '@island.is/web/i18n'
-import NewsListScreen from '@island.is/web/screens/Syslumenn/NewsList'
+import NewsListScreen from '@island.is/web/screens/Organization/NewsList'
 
 export default withApollo(withLocale('is')(NewsListScreen))
