+++ apps/web/pages/stofnanir/syslumenn/uppbod/index.ts
@@ -1,5 +1,5 @@
 import withApollo from '@island.is/web/graphql/withApollo'
 import { withLocale } from '@island.is/web/i18n'
-import Auctions from '@island.is/web/screens/Syslumenn/Auctions'
+import Auctions from '@island.is/web/screens/Organization/Syslumenn/Auctions'
 
 export default withApollo(withLocale('is')(Auctions))
